Заканчивает генерацию разметки похожих элементов.
